chore(pull_over): move common utils
