+++ planning/behavior_path_planner/include/behavior_path_planner/util/pull_over/goal_searcher.hpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_HPP_
 
-#include "behavior_path_planner/scene_module/pull_over/goal_searcher_base.hpp"
 #include "behavior_path_planner/util/occupancy_grid_based_collision_detector/occupancy_grid_based_collision_detector.hpp"
+#include "behavior_path_planner/util/pull_over/goal_searcher_base.hpp"
 
 #include "autoware_auto_planning_msgs/msg/path_point_with_lane_id.hpp"
 
@@ -52,4 +52,4 @@ private:
 };
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_HPP_
