+++ planning/behavior_path_planner/src/util/pull_over/goal_searcher.cpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "behavior_path_planner/scene_module/pull_over/goal_searcher.hpp"
+#include "behavior_path_planner/util/pull_over/goal_searcher.hpp"
 
 #include "behavior_path_planner/path_utilities.hpp"
-#include "behavior_path_planner/scene_module/pull_over/util.hpp"
+#include "behavior_path_planner/util/pull_over/util.hpp"
 #include "lanelet2_extension/utility/utilities.hpp"
 
 #include <boost/optional.hpp>
