+++ planning/behavior_path_planner/src/util/pull_over/util.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "behavior_path_planner/scene_module/pull_over/util.hpp"
+#include "behavior_path_planner/util/pull_over/util.hpp"
 
 #include "behavior_path_planner/path_utilities.hpp"
 #include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
